use mixin
