+++ client/interledger/js/components/interledger.js
@@ -8,33 +8,13 @@ import AccountList from '../../../lib/js/react/components/accounts';
 import AccountDetail from './account_detail';
 
 import AssetActions from '../../../lib/js/react/actions/asset_actions';
-import AssetStore from '../../../lib/js/react/stores/asset_store';
+
+import BigchainDBMixin from '../../../lib/js/react/mixins/bigchaindb_mixin';
+
 
 const Interledger = React.createClass({
 
-    getInitialState() {
-        const assetStore = AssetStore.getState();
-
-        return safeMerge(
-            {
-                activeAccount: null,
-                activeAsset: null
-            },
-            assetStore
-        );
-    },
-
-    componentDidMount() {
-        AssetStore.listen(this.onChange);
-    },
-
-    componentWillUnmount() {
-        AssetStore.unlisten(this.onChange);
-    },
-
-    onChange(state) {
-        this.setState(state);
-    },
+    mixins: [BigchainDBMixin],
 
     fetchAssetList({ accountToFetch }) {
         if (accountToFetch) {
@@ -42,39 +22,6 @@ const Interledger = React.createClass({
                 accountToFetch
             });
         }
-    },
-    
-    handleAccountChange(account) {
-        if (account.ledger) {
-            account.ledger.on('incoming', this.handleLedgerChanges);
-        }
-        this.setState({
-            activeAccount: account
-        });
-
-        const accountToFetch = account ? account.vk : null;
-        this.fetchAssetList({
-            accountToFetch
-        });
-    },
-    
-    resetActiveAccount() {
-        this.handleAccountChange(null, null);
-    },
-
-    handleLedgerChanges(changes) {
-        console.log('incoming: ', changes);
-        const { activeAccount } = this.state;
-
-        this.fetchAssetList({
-            accountToFetch: activeAccount.vk
-        });
-    },
-
-    setActiveAsset(asset) {
-        this.setState({
-            activeAsset: asset
-        });
     },
 
     render() {
